challenge-1/requirements.txt
open-clip-torch>=2.20.0
pandas>=1.5.0
pillow>=9.0.0
scikit-learn>=1.0.0
torch>=1.12.0
numpy>=1.20.0
matplotlib>=3.5.0
